materials_lesson/08.LLM_Memory_FunctionCall/pyproject.toml
[project]
name = "08-llm-memory-functioncall"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
